+++ yash-env/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to `yash-env` will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.1] - Unreleased
+
+### Added
+
+- The `System` trait now has the `getsid` method.
+    - This method returns the session ID of the process with the given PID.
+
 ## [0.6.0] - 2025-03-23
 
 ### Added